remove setupWebConfig as app should query application entity in runtime
