+++ eagle-core/eagle-app/eagle-app-base/src/main/java/org/apache/eagle/app/spi/ApplicationProvider.java
@@ -34,9 +34,4 @@ public interface ApplicationProvider<T extends Application> {
      * @return application instance
      */
     T getApplication();
-
-    /**
-     * set configuration for web resource when application onInstall is invoked
-     */
-    void installWebConfig(Config config);
 }
