+++ eagle-security/eagle-security-common/pom.xml
@@ -52,6 +52,11 @@
       <artifactId>eagle-alert-service</artifactId>
       <version>${project.version}</version>
     </dependency>
+      <dependency>
+          <groupId>org.apache.eagle</groupId>
+          <artifactId>eagle-metadata-base</artifactId>
+          <version>${project.version}</version>
+      </dependency>
 	<dependency>
 		<groupId>org.json</groupId>
 		<artifactId>json</artifactId>
